fixed issue of using doubles in api prompt (api doesn't accept doubles), and fixed formatting for price in meal list adapter

+++ app/src/main/java/com/example/mealpreptracker/Ingredient.kt
@@ -7,7 +7,7 @@ data class Ingredient (
     val id: String = "",
     val meal_id: String? = null,
     var name: String? = null,
-    var quantity: Double? = null,
+    var quantity: Int? = null,
     var price: Double? = null,
     var nutritionSummary: NutritionSummary? = null
 ):  java.io.Serializable
